Refactor project directory organization
