+++ employee-rostering-backend/src/main/java/org/optaweb/employeerostering/service/roster/RosterService.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaweb.employeerostering.service;
+package org.optaweb.employeerostering.service.roster;
 
 import java.util.Optional;
 
@@ -22,7 +22,7 @@ import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
 
 import org.optaweb.employeerostering.domain.roster.RosterState;
-import org.optaweb.employeerostering.persistence.RosterStateRepository;
+import org.optaweb.employeerostering.service.common.AbstractRestService;
 import org.springframework.stereotype.Service;
 
 @Service
