+++ employee-rostering-backend/src/main/java/org/optaweb/employeerostering/service/rotation/RotationController.java
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package org.optaweb.employeerostering.controller;
+package org.optaweb.employeerostering.service.rotation;
 
 import java.util.List;
 
 import org.optaweb.employeerostering.domain.rotation.view.ShiftTemplateView;
-import org.optaweb.employeerostering.service.RotationService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.util.Assert;
@@ -43,7 +42,7 @@ public class RotationController {
         Assert.notNull(rotationService, "rotationService must not be null.");
     }
 
-    @GetMapping
+    @GetMapping("/")
     public ResponseEntity<List<ShiftTemplateView>> getShiftTemplateList(@PathVariable Integer tenantId) {
         return new ResponseEntity<>(rotationService.getShiftTemplateList(tenantId), HttpStatus.OK);
     }
