+++ employee-rostering-backend/src/test/java/org/optaweb/employeerostering/contract/ContractServiceTest.java
@@ -23,7 +23,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import org.optaweb.employeerostering.domain.contract.Contract;
-import org.optaweb.employeerostering.service.ContractService;
+import org.optaweb.employeerostering.service.contract.ContractService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -57,7 +57,7 @@ public class ContractServiceTest {
     @Test
     public void getContractListTest() throws Exception {
         mvc.perform(MockMvcRequestBuilders
-                            .get("/rest/tenant/{tenantId}/contract", 1)
+                            .get("/rest/tenant/{tenantId}/contract/", 1)
                             .accept(MediaType.APPLICATION_JSON))
                 .andDo(mvcResult -> logger.info(mvcResult.toString()))
                 .andExpect(status().isOk());
